+++ packages/typespec-rust/CHANGELOG.md
@@ -12,7 +12,9 @@
 ### Features Added
 
 * Define a `pub(crate)` constant for `api-version` to use in hand-authored `Default` implementation on client options.
-* Add support for TypeSpec `union` types.
+* Added support for the following.
+  * TypeSpec `union` types.
+  * Grouped method parameters via the `@override` decorator.
 
 ### Other Changes